refactor: extract request-resilience and Spotify-read modules
Move repeat_on_request_error + _run_rate_limiter into ratelimit.py, and the
Spotify fetchers (_fetch_all_from_spotify_in_chunks, get_tracks_from_spotify_playlist,
get_followed_artists_from_spotify, get_playlists_from_spotify) into spotify_api.py
(mirroring tidalapi_patch.py on the Tidal side). Both are leaf modules with no
dependency on sync, so spotify_api can import the retry helper without a cycle.
sync.py re-exports all moved names, so existing imports and tests are unchanged.
sync.py drops from ~995 to ~697 lines; no behaviour change.
